+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 
 ## [3.2.0] - Unreleased
 ### Added
+- New plugin `epub`, to export the site into a `.epub` file.
 - New plugin `extract_order`, to extract the `order` property from the filename. (i.e. `23.hello-world.md`)
 - `katex` plugin: add support for `mhchem` extension
 - `katex` plugin: Allow not downloading CSS and font assets for KaTeX plugin [#810]